refactor: move tailwind config into its own package
