+++ tooling/tailwind-config/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["@portofolio/tsconfig/base"],
+  "include": ["."]
+}
